смена на long

+++ src/Main.java
@@ -1,8 +1,10 @@
 public class Main {
     public static void main(String[] args) {
-        int[] sales = new int[]{1, 10, 100, 1000};
+        long[] sales = new long[]{1, 10, 100, 1000};
         SalesManager salesmanager = new SalesManager(sales);
-        int max = salesmanager.max();
+        long max = salesmanager.max();
         System.out.println(max);
     }
+
+
 }
